feat: delete item

+++ src/components/Item/index.jsx
@@ -16,7 +16,11 @@ export default class Item extends Component {
     }
   }
   handleDelete = (id) => {
-    this.props.deleteTodo(id);
+    const isDel = window.confirm('确定删除吗?');
+    if(isDel){
+      this.props.deleteTodo(id);
+    }
+    
   }
   render() {
     const { id, name, done } = this.props;
